SLING-7792 - Resource Resolver should return more than one resolved path if available
Add a new ResourceMapperImpl class that contains the logic for fine-grained access to
resource mappings.

Added a new set of tests that validate that the usage of the various mapping sources
(aliases, /etc/map) is correct.

This change also opens up some methods from the ResourceResolverImpl, but it looks
reasonable as they are not exposed outside the bundle.